--- 8LETTE-TextRPG/IEquipable.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace _8LETTE_TextRPG
-{
-    public interface IEquipable
-    {
-        void Equip();
-
-        void Unequip();
-    }
-}
