--- 8LETTE-TextRPG/IUsable.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace _8LETTE_TextRPG
-{
-    public interface IUsable
-    {
-        void Use();
-    }
-}
