+++ 8LETTE-TextRPG/ItemFolder/IEquipable.cs
@@ -1,0 +1,9 @@
+﻿namespace _8LETTE_TextRPG.ItemFolder
+{
+    public interface IEquipable
+    {
+        void Equip();
+
+        void Unequip();
+    }
+}
