+++ 8LETTE-TextRPG/ItemFolder/IUsable.cs
@@ -1,0 +1,7 @@
+﻿namespace _8LETTE_TextRPG.ItemFolder
+{
+    public interface IUsable
+    {
+        void Use();
+    }
+}
