+++ 8LETTE-TextRPG/MonsterFolder/Monster.cs
@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace _8LETTE_TextRPG
+﻿namespace _8LETTE_TextRPG.MonsterFolder
 {
     public abstract class Monster
     {
